Backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=3.13.0 